PB-164 part 1: relocating imagecheck related logic to another file
